More efficient filtering of watched videos
The app now sends videos loaded on the YouTube search results page through the watched videos stored in the service worker to get back which of the video IDs exist. This is the reverse of getting the watched videos and checking if any of the videos exist in the search results.